- Add automatic splitting of the database on node start
