+++ core/database/core.go
@@ -142,6 +142,11 @@ func provide(c *dig.Container) {
 			}
 		}
 
+		// Check if we need to migrate a legacy database into the split format
+		if err := SplitIntoTangleAndUTXO(deps.DatabasePath); err != nil {
+			CorePlugin.Panic(err)
+		}
+
 		targetEngine, err := database.CheckDatabaseEngine(deps.TangleDatabasePath, true, deps.DatabaseEngine)
 		if err != nil {
 			CorePlugin.Panic(err)
